Fix case when message body is a list
When message body is e.g.
    [
      "software.amazon.payloadoffloading.PayloadS3Pointer",
      {
        "s3BucketName": "my-ssq-bucket",
        "s3Key": "af678f3c-ac46-4539-94d6-eac3c5e60db4"
      }
    ]

message parsing fails due to it not finding the keys "s3BucketName" and
"s3Key" in the body.

This commit changes so that we check if the body is a list, and the first
element of the list is the string
"software.amazon.payloadoffloading.PayloadS3Pointer".

If the body is not a list we fall back to the old behavior.